requirements.txt
google-generativeai>=0.4.1
pydantic>=2.0.0
python-dotenv>=1.0.1
asyncio
